A-TOA/a2a-treasury/requirements.txt
# A2A Treasury Network — Phase 1 + Phase 2
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
redis>=5.0.0
pyjwt>=2.8.0
bcrypt>=4.1.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
httpx>=0.27.0
rich>=13.7.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
python-multipart>=0.0.9
# Phase 2
py-algorand-sdk>=2.6.0
algorand-python>=2.0.0
groq>=0.9.0
tenacity>=8.2.0
aiofiles>=23.2.0
# Phase 3
jinja2>=3.1.3
aiohttp>=3.9.0
# Phase 4
gunicorn==21.2.0
# Phase 7: UI
sse-starlette>=1.6.5
itsdangerous>=2.1.2
